api/requirements.txt
a2wsgi==1.8.0
alembic==1.13.1
anyio==4.0.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
connexion==3.0.1
coverage==7.3.2
cryptography==43.0.1
Deprecated==1.2.14
dparse==0.6.3
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Alembic==2.0.1
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.1
idna==3.7
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Mako==1.3.0
MarkupSafe==2.1.3
mysqlclient==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-instrumentation==0.43b0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==21.3
pluggy==1.3.0
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.20.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.18
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.32.0
requests-oauthlib==1.3.1
rpds-py==0.13.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safety==2.3.5
sentry-sdk==2.8.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.40.0
swagger_ui_bundle==1.1.0
tomli==2.0.1
typing_extensions==4.8.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.6
wrapt==1.16.0
zero-totp-db-model==1.12
zipp==3.19.1
